Fix effort #2
On further testing, it was revealed that the DHCP service fails because DHCPDISCOVER messages are not received by the pihole device.

This indicates an issue in L2 broadcast rx. After random testing, I think the wifi gets disconnected due to channel changing and when connecting, sometimes there is a corrupt beacon. I am looking for those in the last five minutes and then bouncing the wifi.